Add flavor text and ai model
